similar error fixed in the previous commit

+++ pre_process/data_loader.py
@@ -14,7 +14,7 @@ def embed_arr(arr, dims):
 
 
 def create_mask(arr_dims, dims):
-    mask = np.zeros(dims)
+    mask = np.zeros(dims, dtype=np.float16)
     mask[:arr_dims[0],:] = 1
     return mask
